chore: replace flake-utils with flake-parts and update flake configuration
This commit replaces the `flake-utils` dependency with `flake-parts` and
updates the flake configuration to use the new `flake-parts` structure.
The `utils.lib.eachDefaultSystem` function is replaced with
`parts.lib.mkFlake`, and the `systems` list is explicitly defined.
Additionally, the `deps` list is removed, and build inputs are directly
specified in the `devShells.default` configuration.

Signed-off-by: Qiming Chu <cchuqiming@gmail.com>